requirements.txt
torch~=2.4.1+cu118
torchvision~=0.19.1+cu118
pandas~=2.0.3
Pillow~=10.2.0
matplotlib~=3.7.5
tqdm~=4.66.5
scikit-learn~=1.3.2
numpy~=1.24.1
scipy~=1.10.1
opencv-python-headless